ue4mtc: Added material & explosion to tracker bot
Added a dynamic material instance to the tracker bot to allow for per object variables. This will allow us to dynamically flash the tracker bot on damage taken.

Added self destruction to the tracker bot, this will destroy the bot upon health is 0 or lower. An effect emitter will be displayed at the location if one is set.